backend/requirements.txt
firebase-admin==6.4.0
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.5.2
python-dotenv==1.0.1
